Fixed linux install prefixes and --untracked option handling (#100)
* Fixed default install prefixes on linux

The bash_completion script and man file were being installed to the
wrong location by default.  They are now installed correctly to
'/etc/bash_completion.d' and '/usr/share/man/man1'.

* Fixed --untracked option handling

There were a couple places where this was handled inconsistently.  First,
the documentation said this is an option for the diff command, but it
isn't.  Second, the bash_completion script didn't autocomplete this
option for the info command.  And third, the -u shorthand was present on
the info command but not the status command.  All of these are now
fixed.